ready

+++ servlet-jsp/multi-api/src/main/java/multi/api/contract/ProductApi.java
@@ -3,11 +3,15 @@ package multi.api.contract;
 import multi.api.dto.product.ProductRequest;
 import multi.api.dto.product.ProductResponse;
 
+
 import java.util.List;
 
 public interface ProductApi {
     void createProduct(ProductRequest productRequest);
+
     void updateProduct(long id, ProductRequest productRequest);
-    void deleteProduct(int id);
+
+    void deleteProduct(long id);
+
     List<ProductResponse> showProduct();
 }
